Refactor
bench 1666201
