+++ Serendipity/src/main/java/org/shawn/games/Serendipity/PeSTO/PeSTO.java
@@ -1,4 +1,4 @@
-package org.shawn.games.Serendipity;
+package org.shawn.games.Serendipity.PeSTO;
 
 import com.github.bhlangonijr.chesslib.*;
