+++ Serendipity/src/main/java/org/shawn/games/Serendipity/Search/Limits.java
@@ -1,4 +1,4 @@
-package org.shawn.games.Serendipity;
+package org.shawn.games.Serendipity.Search;
 
 public class Limits
 {
