+++ Serendipity/src/main/java/org/shawn/games/Serendipity/Search/MoveSort.java
@@ -1,6 +1,8 @@
-package org.shawn.games.Serendipity;
+package org.shawn.games.Serendipity.Search;
 
 import java.util.*;
+
+import org.shawn.games.Serendipity.History.History;
 
 import com.github.bhlangonijr.chesslib.*;
 import com.github.bhlangonijr.chesslib.move.*;
