+++ Serendipity/src/main/java/org/shawn/games/Serendipity/Search/ScoredMove.java
@@ -1,4 +1,4 @@
-package org.shawn.games.Serendipity;
+package org.shawn.games.Serendipity.Search;
 
 import com.github.bhlangonijr.chesslib.move.Move;
