+++ Serendipity/src/main/java/org/shawn/games/Serendipity/Search/SearchStack.java
@@ -1,4 +1,7 @@
-package org.shawn.games.Serendipity;
+package org.shawn.games.Serendipity.Search;
+
+import org.shawn.games.Serendipity.History.History;
+import org.shawn.games.Serendipity.History.PieceToHistory;
 
 import com.github.bhlangonijr.chesslib.move.Move;
