+++ Serendipity/src/main/java/org/shawn/games/Serendipity/UCI/Tunable.java
@@ -1,4 +1,4 @@
-package org.shawn.games.Serendipity;
+package org.shawn.games.Serendipity.UCI;
 
 public class Tunable extends IntegerOption
 {
